latest version, back to numpy for matrix but no glFlush or glFinish

+++ pi3d/Camera.py
@@ -33,7 +33,7 @@ class Camera(object):
     return copy(self.mtrx)
 
   def identity(self):
-    self.mtrx = array([[1,0,0,0],[0,1,0,0],[0,0,1,0],[0,0,0,1]])
+    self.mtrx = array([[1,0,0,0],[0,1,0,0],[0,0,1,0],[0,0,0,1]], dtype=c_float)
     
   def translate(self,pt):
     #self.mtrx = translate(self.mtrx,pt)
@@ -87,7 +87,7 @@ def LookAtMatrix(at, eye, up=[0,1,0], reflect=False):
   zaxis.append(-vec_dot(zaxis,eye))
   z = [0,0,0,1.0]
   #TODO all using numpy functions
-  return array([[xaxis[a],yaxis[a],zaxis[a],z[a]] for a in range(4)])
+  return array([[xaxis[a],yaxis[a],zaxis[a],z[a]] for a in range(4)], dtype=c_float)
   #return array([xaxis,yaxis,zaxis,z])
 
 def ProjectionMatrix(near=10, far=1000.0, fov_h=1.6, fov_v=1.2):
@@ -107,5 +107,5 @@ def ProjectionMatrix(near=10, far=1000.0, fov_h=1.6, fov_v=1.2):
   #M[2][3] = -Q*near
   #M[3][2] = 1
   #TODO all using numpy functions
-  return array(M)
+  return array(M, dtype=c_float)
